backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.2
django-autoslug==1.9.9
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.1
Faker==23.3.0
idna==3.6
inflection==0.5.1
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
gunicorn==20.1.0
psycopg2-binary
